PngDecoderTests.Chunks.cs
